+++ app/src/main/java/org/simple/clinic/AppDatabase.kt
@@ -39,8 +39,6 @@ import org.simple.clinic.summary.teleconsultation.sync.TeleconsultationFacilityW
 import org.simple.clinic.teleconsultlog.medicinefrequency.MedicineFrequency
 import org.simple.clinic.teleconsultlog.teleconsultrecord.Answer.RoomTypeConverter
 import org.simple.clinic.teleconsultlog.teleconsultrecord.TeleconsultRecord
-import org.simple.clinic.teleconsultlog.teleconsultrecord.TeleconsultRecordPrescribedDrug
-import org.simple.clinic.teleconsultlog.teleconsultrecord.TeleconsultRecordWithPrescribedDrugs
 import org.simple.clinic.teleconsultlog.teleconsultrecord.TeleconsultationType
 import org.simple.clinic.user.OngoingLoginEntry
 import org.simple.clinic.user.User
@@ -70,14 +68,13 @@ import org.simple.clinic.util.room.UuidRoomTypeConverter
       TeleconsultationFacilityInfo::class,
       MedicalOfficer::class,
       TeleconsultationFacilityMedicalOfficersCrossRef::class,
-      TeleconsultRecord::class,
-      TeleconsultRecordPrescribedDrug::class
+      TeleconsultRecord::class
     ],
     views = [
       OverdueAppointment::class,
       PatientSearchResult::class
     ],
-    version = 75,
+    version = 76,
     exportSchema = true
 )
 @TypeConverters(
@@ -153,10 +150,6 @@ abstract class AppDatabase : RoomDatabase() {
 
   abstract fun teleconsultRecordDao(): TeleconsultRecord.RoomDao
 
-  abstract fun teleconsultRecordPrescribedDrugDao(): TeleconsultRecordPrescribedDrug.RoomDao
-
-  abstract fun teleconsultRecordWithPrescribedDrugDao(): TeleconsultRecordWithPrescribedDrugs.RoomDao
-
   fun clearAppData() {
     runInTransaction {
       patientDao().clear()
@@ -171,7 +164,6 @@ abstract class AppDatabase : RoomDatabase() {
       teleconsultMedicalOfficersDao().clear()
       teleconsultFacilityWithMedicalOfficersDao().clear()
       teleconsultRecordDao().clear()
-      teleconsultRecordPrescribedDrugDao().clear()
     }
   }