+++ app/src/main/java/org/simple/clinic/teleconsultlog/teleconsultrecord/TeleconsultRecord.kt
@@ -38,5 +38,10 @@ data class TeleconsultRecord(
     @Query("DELETE FROM TeleconsultRecord")
     fun clear()
 
+    @Query("SELECT * FROM TeleconsultRecord")
+    fun getAll(): List<TeleconsultRecord>
+
+    @Query("SELECT * FROM TeleconsultRecord WHERE id = :teleconsultRecordId")
+    fun getCompleteTeleconsultLog(teleconsultRecordId: UUID) : TeleconsultRecord
   }
 }